two separate scripts for flight planning and scene writing, fixed imports, adapted tests
